second

+++ src/main/java/com/tradeone/model/Rialto.java
@@ -7,6 +7,8 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 import lombok.Data;
 
@@ -15,17 +17,20 @@ import lombok.Data;
 @Data
 public class Rialto {
 
+
   @Id
   @GeneratedValue(strategy = IDENTITY)
   private Long id;
   @Column(name = "name", nullable = false)
   private String name;
   @Column(name = "tradingBalance", nullable = false)
-  private long balance;
+  private double balance ;
   @Column(name = "profit", nullable = false)
   private int profit;
   @Column(name = "ordersByDay", nullable = false)
   private int ordersByDay;
 
 
+
 }
+
